+++ llm_chat/domain/usecase/completion/chat.py
@@ -7,7 +7,7 @@ from lib.llm.valueobject.completion import RoleType, StreamResponse
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
 from llm_chat.domain.service.completion.chat import ChatService, OpenAIChatStreamingService
 from llm_chat.domain.usecase.completion.base import UseCase
-from llm_chat.domain.valueobject.completion.message import MessageDTO
+from llm_chat.domain.valueobject.completion.chat import MessageDTO
 
 
 class LlmChatUseCase(UseCase):
