+++ llm_chat/domain/valueobject/completion/chat.py
