+++ pr_body.txt
@@ -1,0 +1,54 @@
+## 概要
+llm_chat/domain 配下の各ディレクトリにおいて、chat.py に集約されていたコードを責務ごとに分割し、構造を整理しました。
+
+## 変更内容
+
+### 1. 責務ごとにファイルを分割
+- **service/chat.py** を分割：
+  - base.py: BaseChatService
+  - chat.py: ChatService, OpenAIChatStreamingService
+  - riddle.py: RiddleTask、なぞなぞ機能
+  - multimedia.py: DALL-E, TTS, STT
+  - rag.py: RAG機能
+
+- **usecase/chat.py** を分割：
+  - base.py: UseCase基底クラス
+  - chat.py: LlmChatUseCase, OpenAIGptStreamingUseCase
+  - riddle.py: RiddleUseCase
+  - multimedia.py: OpenAIDalleUseCase, OpenAITextToSpeechUseCase, OpenAISpeechToTextUseCase
+  - rag.py: OpenAIRagUseCase
+
+### 2. completion/ サブディレクトリに整理
+将来的な拡張性を考慮し、completion/ サブディレクトリを作成：
+- repository/completion/
+- service/completion/
+- usecase/completion/
+- valueobject/completion/
+
+### 3. valueobject も分割、全 __init__.py をカラに
+- **valueobject/chat.py** を分割：
+  - completion/message.py: MessageDTO
+  - completion/riddle.py: Gender, GenderType, RiddleEvaluation, RiddleResponse
+- すべての __init__.py をカラにして明示的なインポートを強制
+
+### 4. ファイル命名の統一
+他のファイル（rag.py, multimedia.py, riddle.py）が機能名で命名されているため、統一感を持たせるために common.py を chat.py にリネーム
+
+## 最終的なディレクトリ構造
+
+
+## テスト観点 (オペレーション視点)
+- [x] 既存のテストがすべてパス (12 tests OK)
+- [x] インポート文が正しく更新されている
+- [x] 各機能が正常に動作する（Chat, Riddle, Multimedia, RAG）
+
+### 自動テストでカバーできた範囲
+- 通常チャット機能
+- なぞなぞ機能（評価含む）
+- DALL-E画像生成
+- TTS音声生成
+- STT音声認識
+- RAG機能
+
+## 関連Issue
+Closes #583
